requirements.txt
# Core dependencies
solana>=0.30.2
solders>=0.18.1
anchorpy>=0.18.0
base58>=2.1.1
construct>=2.10.68

# Web and networking
websockets>=12.0
aiohttp>=3.9.1
requests>=2.31.0

# Data handling
pyyaml>=6.0.1
python-dotenv>=1.0.0
pandas>=2.1.4

# Cryptography
cryptography>=41.0.7
pynacl>=1.5.0

# Utilities
asyncio>=3.4.3
typing-extensions>=4.9.0
python-dateutil>=2.8.2

# Logging and monitoring
colorlog>=6.8.0
rich>=13.7.0

